# swagger.yaml
openapi: 3.1.0
info:
  title: Trading Platform Backend API
  description: |
    A comprehensive trading platform backend with IIFL integration, real-time market data, 
    portfolio management, and advanced stock analysis features.
    
    ## Features
    - **Google OAuth Authentication** - Secure user authentication
    - **IIFL Trading Integration** - Place orders, manage positions
    - **Real-time Market Data** - Live prices, quotes, and market analytics
    - **Portfolio Management** - Track P&L, holdings, risk metrics
    - **Stock Analysis** - Advanced analytics and returns calculation
    - **WebSocket Support** - Real-time updates
    
  version: 1.0.0
  contact:
    name: Trading Platform Support
    email: support@tradingplatform.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://trading-backend-oab.info
    description: Production server
...
paths:
...
  /api/portfolio/summary:
    get:
      summary: Get portfolio summary
      description: Get comprehensive portfolio summary with positions, P&L, and performance metrics
      tags:
        - Portfolio
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Portfolio summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
  /api/portfolio/pnl:
    get:
      summary: Get P&L calculation
      description: Calculate profit and loss for a specified date range
      tags:
        - Portfolio
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Start date for P&L calculation
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: End date for P&L calculation
      responses:
        '200':
          description: P&L calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PnLData'
  /api/portfolio/update-prices:
    post:
      summary: Update position prices
      description: Update current market prices for all positions in the portfolio
      tags:
        - Portfolio
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Prices updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_count:
                    type: integer
                  message:
                    type: string
...
  /api/portfolio/daily-pnl:
    get:
      summary: Get today's P&L
      description: Get profit and loss for the current trading day
      tags:
        - Portfolio
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daily P&L retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  daily_pnl:
                    type: number
                  realized_pnl:
                    type: number
                  unrealized_pnl:
                    type: number
  /api/portfolio/holdings:
    get:
      summary: Get long-term holdings
      description: Retrieve user's long-term stock holdings from IIFL
      tags:
        - Portfolio
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Holdings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  holdings:
                    type: object
                  message:
                    type: string
  /api/portfolio/holdings-summary:
    get:
      summary: Get holdings summary with live prices
      description: Get user-friendly holdings summary from IIFL with current market prices
      tags:
        - Portfolio
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Holdings summary retrieved successfully

  # ========== MARKET DATA ENDPOINTS ==========
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from Google OAuth login.
        
        **How to authenticate:**
        1. Login via `/api/auth/oauth/google/login`
        2. Complete OAuth flow
        3. Receive `access_token` and `refresh_token`
        4. Use `access_token` in Authorization header: `Bearer {token}`
        5. Refresh token when expired using `/api/auth/refresh`
  schemas:
...
    PortfolioSummary:
      type: object
      properties:
        total_positions:
          type: integer
        long_positions:
          type: integer
        short_positions:
          type: integer
        total_investment:
          type: number
        current_value:
          type: number
        unrealized_pnl:
          type: number
        daily_pnl:
          type: number
        monthly_pnl:
          type: number
        positions:
          type: array
          items:
            type: object
    PnLData:
      type: object
      properties:
        total_realized_pnl:
          type: number
        total_unrealized_pnl:
          type: number
        total_pnl:
          type: number
        total_trades:
          type: integer
        winning_trades:
          type: integer
        losing_trades:
          type: integer
        win_rate:
          type: number
        total_charges:
          type: number
...
security:
  - BearerAuth: []